+++ packages/devui-vue/devui/form/src/form.tsx
@@ -1,10 +1,8 @@
-
 import { defineComponent, provide } from 'vue';
 import mitt from 'mitt';
 import { formProps, FormProps, IFormItem, dFormEvents, formInjectionKey, IForm } from './form-types';
 import { EventBus } from './util';
 import './form.scss';
-
 
 export default defineComponent({
   name: 'DForm',
@@ -12,7 +10,7 @@ export default defineComponent({
   emits: ['submit'],
   setup(props: FormProps, ctx) {
     const formMitt = mitt();
-    const fields: IFormItem[] =  [];
+    const fields: IFormItem[] = [];
     const resetFormFields = () => {
       fields.forEach((field: IFormItem) => {
         field.resetField();
@@ -20,13 +18,13 @@ export default defineComponent({
     };
 
     formMitt.on(dFormEvents.addField, (field: any) => {
-      if(field) {
+      if (field) {
         fields.push(field);
       }
     });
 
     formMitt.on(dFormEvents.removeField, (field: any) => {
-      if(field.prop) {
+      if (field.prop) {
         fields.splice(fields.indexOf(field), 1);
       }
     });
@@ -41,7 +39,7 @@ export default defineComponent({
       },
       rules: props.rules,
       columnsClass: props.columnsClass,
-      messageShowType: 'popover'
+      messageShowType: 'popover',
     });
 
     const onSubmit = (e) => {
@@ -54,15 +52,15 @@ export default defineComponent({
       fields,
       formMitt,
       onSubmit,
-      resetFormFields
+      resetFormFields,
     };
   },
   render() {
-    const {onSubmit} = this;
+    const { onSubmit } = this;
     return (
       <form onSubmit={onSubmit} class="devui-form">
         {this.$slots.default?.()}
       </form>
     );
-  }
+  },
 });
